Move Formatters to Outputs
